requirements.txt
streamlit
PyPDF2
python-docx
python-dotenv
google-genai